# specs/001-compensation-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Maliev Compensation Service API
  description: |
    Employee Compensation Management Service API for managing compensation records,
    salary history, benefits enrollment, and bulk operations.

    **Security**: All endpoints require JWT authentication and permission-based authorization.
    All operations are marked as critical and generate audit log entries.
  version: 1.0.0
  contact:
    name: MALIEV Core Developers
    email: support@maliev.com
servers:
  - url: https://api.maliev.com/compensation/v1
    description: Production
  - url: https://staging-api.maliev.com/compensation/v1
    description: Staging
  - url: http://localhost:8080/compensation/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /benefits:
    get:
      tags: [Benefits]
      summary: List available benefits
      description: Retrieves all active benefits programs
      operationId: getAvailableBenefits
      parameters:
        - name: benefitType
          in: query
          schema:
            $ref: '#/components/schemas/BenefitType'
          description: Filter by benefit type
      responses:
        '200':
          description: Benefits list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BenefitDto'
      security:
        - BearerAuth: [compensation.read]
  /employees/{employeeId}/benefits:
    get:
      tags: [Benefits]
      summary: Get employee benefits
      description: Retrieves all benefit enrollments for an employee
      operationId: getEmployeeBenefits
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EnrollmentStatus'
          description: Filter by enrollment status
      responses:
        '200':
          description: Employee benefits retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BenefitsEnrollmentDto'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [compensation.read]

    put:
      tags: [Benefits]
      summary: Update benefits enrollment
      description: Updates an existing benefits enrollment (add dependents, change contribution)
      operationId: updateBenefitsEnrollment
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBenefitsEnrollmentDto'
      responses:
        '200':
          description: Benefits enrollment updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [compensation.update]
  /employees/{employeeId}/benefits/{benefitId}/enroll:
    post:
      tags: [Benefits]
      summary: Enroll in benefit
      description: Enrolls an employee in a benefit program
      operationId: enrollInBenefit
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
        - name: benefitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [coverageLevel]
              properties:
                coverageLevel:
                  type: string
                  enum: [Individual, Family, Spouse, Spouse+Children]
                employeeContribution:
                  type: number
                  format: decimal
                dependents:
                  type: array
                  items:
                    $ref: '#/components/schemas/DependentDto'
      responses:
        '201':
          description: Enrollment created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
      security:
        - BearerAuth: [compensation.update]
  /employees/{employeeId}/benefits/{benefitId}:
    delete:
      tags: [Benefits]
      summary: Terminate benefit
      description: Terminates an active benefits enrollment
      operationId: terminateBenefit
      parameters:
        - $ref: '#/components/parameters/EmployeeIdPath'
        - name: benefitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: terminationDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '204':
          description: Benefit terminated successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [compensation.update]

  # Bulk Operations Endpoints
  /bulk/salary-increase:
    post:
      tags: [Bulk Operations]
      summary: Apply bulk salary increase
      description: Applies salary increase to multiple employees (max 500 per job)
      operationId: bulkSalaryIncrease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkSalaryIncreaseDto'
      responses:
        '202':
          description: Bulk operation accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkJobStatusDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - BearerAuth: [compensation.admin]
  /bulk/jobs/{jobId}:
    get:
      tags: [Bulk Operations]
      summary: Get bulk job status
      description: Retrieves the status of a bulk operation job
      operationId: getBulkJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkJobStatusDto'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - BearerAuth: [compensation.admin]

  # Reports Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EmployeeIdPath:
      name: employeeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Employee unique identifier
  schemas:
...
    BenefitDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        benefitType:
          $ref: '#/components/schemas/BenefitType'
        employerContribution:
          type: number
          format: decimal
        employeeContribution:
          type: number
          format: decimal
    BenefitsEnrollmentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        benefitId:
          type: string
          format: uuid
        enrollmentDate:
          type: string
          format: date-time
        terminationDate:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/EnrollmentStatus'
        employeeContribution:
          type: number
          format: decimal
        coverageLevel:
          type: string
        dependents:
          type: array
          items:
            $ref: '#/components/schemas/DependentDto'
    UpdateBenefitsEnrollmentDto:
      type: object
      required: [benefitId]
      properties:
        benefitId:
          type: string
          format: uuid
        coverageLevel:
          type: string
        employeeContribution:
          type: number
          format: decimal
        dependents:
          type: array
          items:
            $ref: '#/components/schemas/DependentDto'
    DependentDto:
      type: object
      required: [firstName, lastName, relationship, dateOfBirth]
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        relationship:
          $ref: '#/components/schemas/DependentRelationship'
        dateOfBirth:
          type: string
          format: date
        nationalId:
          type: string
          description: "[SENSITIVE - Encrypted]"
    BulkSalaryIncreaseDto:
      type: object
      required: [departmentId, increasePercentage, effectiveDate, reason]
      properties:
        departmentId:
          type: string
          format: uuid
        increasePercentage:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        effectiveDate:
          type: string
          format: date-time
        reason:
          type: string
          maxLength: 500
        webhookUrl:
          type: string
          format: uri
          description: Optional callback URL for job completion
    BulkJobStatusDto:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [Pending, InProgress, Completed, Failed, PartiallyCompleted]
        affectedEmployees:
          type: integer
        successCount:
          type: integer
        failureCount:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    BenefitType:
      type: string
      enum: [HealthInsurance, DentalInsurance, VisionInsurance, LifeInsurance, Retirement401k, StockOptions, PaidTimeOff, WellnessProgram, EducationAssistance]
    EnrollmentStatus:
      type: string
      enum: [Active, Pending, Terminated, OnHold]
    DependentRelationship:
      type: string
      enum: [Spouse, Child, DomesticPartner, Parent, Other]

    # Error Responses
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidSalary:
              value:
                code: INVALID_SALARY
                message: Salary must be positive
            invalidCurrency:
              value:
                code: INVALID_CURRENCY
                message: Invalid currency code
            invalidEffectiveDate:
              value:
                code: INVALID_EFFECTIVE_DATE
                message: Date must be current or future
            changeLimitExceeded:
              value:
                code: CHANGE_LIMIT_EXCEEDED
                message: Increase exceeds 25% limit
    Forbidden:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_AUTHORIZED
            message: User cannot access compensation
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            employeeNotFound:
              value:
                code: EMPLOYEE_NOT_FOUND
                message: Employee not found
            benefitNotFound:
              value:
                code: BENEFIT_NOT_FOUND
                message: Benefit not found
    Conflict:
      description: Conflict - duplicate or concurrent update
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            enrollmentExists:
              value:
                code: ENROLLMENT_EXISTS
                message: Already enrolled in benefit
            concurrentUpdate:
              value:
                code: CONCURRENT_UPDATE
                message: Record was modified by another user, please refresh and retry